+++ src/pages/NowPlaying.tsx
@@ -32,91 +32,15 @@ import {
 
 // ─── Helpers ──────────────────────────────────────────────────────────────────
 
-function formatTime(s: number): string {
-  if (!s || isNaN(s)) return '0:00';
-  const m = Math.floor(s / 60);
-  return `${m}:${Math.floor(s % 60).toString().padStart(2, '0')}`;
-}
-
-function formatCompact(n: number): string {
-  if (n >= 1_000_000) return `${(n / 1_000_000).toFixed(n >= 10_000_000 ? 0 : 1)}M`;
-  if (n >= 1_000)     return `${(n / 1_000).toFixed(n >= 10_000 ? 0 : 1)}K`;
-  return String(n);
-}
-
-function formatTotalDuration(s: number): string {
-  if (!s || isNaN(s)) return '—';
-  const h = Math.floor(s / 3600);
-  const m = Math.floor((s % 3600) / 60);
-  const sec = Math.floor(s % 60);
-  if (h > 0) return `${h}h ${m}m`;
-  if (m > 0) return `${m}m ${sec}s`;
-  return `${sec}s`;
-}
-
-function sanitizeHtml(html: string): string {
-  const parser = new DOMParser();
-  const doc = parser.parseFromString(html, 'text/html');
-  doc.querySelectorAll('script, style, iframe, object, embed, form, input, button, select, base, meta, link').forEach(el => el.remove());
-  doc.querySelectorAll('*').forEach(el => {
-    Array.from(el.attributes).forEach(attr => {
-      const name = attr.name.toLowerCase();
-      const val = attr.value.toLowerCase().trim();
-      if (name.startsWith('on') || (name === 'href' && (val.startsWith('javascript:') || val.startsWith('data:'))) || (name === 'src' && (val.startsWith('javascript:') || val.startsWith('data:')))) {
-        el.removeAttribute(attr.name);
-      }
-    });
-  });
-  // Strip trailing "Read more on Last.fm" style links for cleaner clamped bios.
-  return doc.body.innerHTML.replace(/<a [^>]*>.*?<\/a>\.?\s*$/i, '').trim();
-}
-
-function isoToParts(iso: string): { month: string; day: string; weekday: string; time: string } | null {
-  if (!iso) return null;
-  const d = new Date(iso);
-  if (Number.isNaN(d.getTime())) return null;
-  return {
-    month: d.toLocaleString(undefined, { month: 'short' }),
-    day: String(d.getDate()),
-    weekday: d.toLocaleString(undefined, { weekday: 'short' }),
-    time: d.toLocaleString(undefined, { hour: '2-digit', minute: '2-digit' }),
-  };
-}
-
-interface ContributorRow { role: string; names: string[]; }
-
-function buildContributorRows(song: SubsonicSong | null | undefined, mainArtistName: string): ContributorRow[] {
-  if (!song?.contributors || song.contributors.length === 0) return [];
-  const mainLower = mainArtistName.trim().toLowerCase();
-  const rows = new Map<string, Set<string>>();
-  for (const c of song.contributors) {
-    const role = c.role?.trim();
-    const name = c.artist?.name?.trim();
-    if (!role || !name) continue;
-    const label = c.subRole ? `${role} • ${c.subRole}` : role;
-    let bucket = rows.get(label);
-    if (!bucket) { bucket = new Set(); rows.set(label, bucket); }
-    bucket.add(name);
-  }
-  const out: ContributorRow[] = [];
-  for (const [role, names] of rows.entries()) {
-    const list = Array.from(names);
-    if (role.toLowerCase().startsWith('artist') && list.length === 1 && list[0].toLowerCase() === mainLower) continue;
-    out.push({ role, names: list });
-  }
-  return out;
-}
-
-/**
- * Filter out the well-known Last.fm "no image" placeholder that Subsonic
- * backends aggregate into `largeImageUrl`/`mediumImageUrl` when no real
- * artist image exists. The placeholder MD5 is fixed and documented.
- */
-function isRealArtistImage(url?: string): boolean {
-  if (!url) return false;
-  if (url.includes('2a96cbd8b46e442fc41c2b86b821562f')) return false;
-  return true;
-}
+import {
+  formatTime, formatCompact, formatTotalDuration, sanitizeHtml, isoToParts,
+  buildContributorRows, isRealArtistImage,
+  type ContributorRow,
+} from '../utils/nowPlayingHelpers';
+import { makeCache } from '../utils/nowPlayingCache';
+import NpCardWrap from '../components/nowPlaying/NpCardWrap';
+import NpColumnEl from '../components/nowPlaying/NpColumnEl';
+import RadioView from '../components/nowPlaying/RadioView';
 
 function renderStars(rating?: number) {
   if (!rating) return null;
@@ -133,23 +57,6 @@ function renderStars(rating?: number) {
 }
 
 // ─── Module-level TTL caches (shared across mounts) ───────────────────────────
-
-const CACHE_TTL_MS = 5 * 60 * 1000;
-
-interface CacheEntry<T> { value: T; ts: number; }
-
-function makeCache<T>() {
-  const map = new Map<string, CacheEntry<T>>();
-  return {
-    get(key: string): T | undefined {
-      const e = map.get(key);
-      if (!e) return undefined;
-      if (Date.now() - e.ts > CACHE_TTL_MS) { map.delete(key); return undefined; }
-      return e.value;
-    },
-    set(key: string, value: T) { map.set(key, { value, ts: Date.now() }); },
-  };
-}
 
 const songMetaCache    = makeCache<SubsonicSong | null>();
 const artistInfoCache  = makeCache<SubsonicArtistInfo | null>();
@@ -718,178 +625,6 @@ const DiscographyCard = memo(function DiscographyCard({ artistId, albums, curren
             ? t('nowPlaying.showLessTracks', 'Show less')
             : t('nowPlaying.showMoreTracks', { defaultValue: 'Show {{count}} more', count: hiddenCount })}
         </button>
-      )}
-    </div>
-  );
-});
-
-// ─── Widget wrapper (drag source via psyDnD) ─────────────────────────────────
-
-interface NpCardWrapProps {
-  id: NpCardId;
-  label: string;
-  isDraggingThis: boolean;
-  children: React.ReactNode;
-}
-
-function NpCardWrap({ id, label, isDraggingThis, children }: NpCardWrapProps) {
-  const dragProps = useDragSource(() => ({
-    data: JSON.stringify({ kind: 'np-card', id }),
-    label,
-  }));
-  return (
-    <div
-      data-np-wrapper
-      data-np-card-id={id}
-      className={`np-dash-card-wrap${isDraggingThis ? ' is-dragging' : ''}`}
-      {...dragProps}
-    >
-      {children}
-    </div>
-  );
-}
-
-// ─── Column (drop target via psy-drop + global mousemove) ────────────────────
-
-interface NpColumnProps {
-  col: NpColumn;
-  children: React.ReactNode;
-  empty: boolean;
-  emptyLabel: string;
-  isDndActive: boolean;
-  draggingCardId: NpCardId | null;
-  onHover: (col: NpColumn, idx: number) => void;
-  isOverHere: boolean;
-}
-
-function NpColumnEl({ col, children, empty, emptyLabel, isDndActive, draggingCardId, onHover, isOverHere }: NpColumnProps) {
-  const ref = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    if (!draggingCardId) return;
-    const el = ref.current;
-    if (!el) return;
-
-    const onMove = (e: MouseEvent) => {
-      const rect = el.getBoundingClientRect();
-      // Use only the x-axis to decide "which column". This keeps the whole
-      // vertical strip above / below the last card part of the drop zone,
-      // so the user can drop "at the very bottom" of either column.
-      if (e.clientX < rect.left || e.clientX > rect.right) return;
-      const wrappers = Array.from(el.querySelectorAll<HTMLElement>('[data-np-wrapper]'))
-        .filter(w => w.getAttribute('data-np-card-id') !== draggingCardId);
-      let idx = wrappers.length;
-      for (let i = 0; i < wrappers.length; i++) {
-        const r = wrappers[i].getBoundingClientRect();
-        if (e.clientY < r.top + r.height / 2) { idx = i; break; }
-      }
-      onHover(col, idx);
-    };
-
-    document.addEventListener('mousemove', onMove);
-    return () => document.removeEventListener('mousemove', onMove);
-  }, [draggingCardId, col, onHover]);
-
-  return (
-    <div
-      ref={ref}
-      className={`np-dash-col${isOverHere ? ' is-drop-target' : ''}${isDndActive ? ' is-dnd-active' : ''}`}
-    >
-      {children}
-      {empty && <div className="np-dash-col-empty">{emptyLabel}</div>}
-    </div>
-  );
-}
-
-type NonNullStoreField<K extends keyof ReturnType<typeof usePlayerStore.getState>> =
-  NonNullable<ReturnType<typeof usePlayerStore.getState>[K]>;
-
-interface RadioViewProps {
-  radioMeta: ReturnType<typeof useRadioMetadata>;
-  currentRadio: NonNullStoreField<'currentRadio'>;
-  resolvedCover: string;
-}
-
-const RadioView = memo(function RadioView({ radioMeta, currentRadio, resolvedCover }: RadioViewProps) {
-  const { t } = useTranslation();
-  return (
-    <div className="np-radio-section">
-      <div className="np-hero-card">
-        <div className="np-hero-left">
-          <div className="np-hero-info">
-            <div className="np-title" style={{ color: 'var(--accent)' }}>{currentRadio.name}</div>
-            {radioMeta.currentTitle && (
-              <div className="np-artist-album">
-                {radioMeta.currentArtist && (<><span className="np-link">{radioMeta.currentArtist}</span><span className="np-sep">·</span></>)}
-                <span>{radioMeta.currentTitle}</span>
-                {radioMeta.currentAlbum && (<><span className="np-sep">·</span><span style={{ opacity: 0.6 }}>{radioMeta.currentAlbum}</span></>)}
-              </div>
-            )}
-            <div className="np-tech-row">
-              <span className="np-badge np-badge-live"><Radio size={10} style={{ marginRight: 3 }} />{t('radio.live')}</span>
-              {radioMeta.source === 'azuracast' && <span className="np-badge np-badge-azuracast">AzuraCast</span>}
-              {radioMeta.listeners != null && (
-                <span className="np-badge"><Users size={10} style={{ marginRight: 3 }} />{t('radio.listenerCount', { count: radioMeta.listeners })}</span>
-              )}
-            </div>
-            {radioMeta.source === 'azuracast' && radioMeta.elapsed != null && radioMeta.duration != null && radioMeta.duration > 0 && (
-              <div className="np-radio-progress-wrap">
-                <span className="np-radio-time">{formatTime(radioMeta.elapsed)}</span>
-                <div className="np-radio-progress-bar">
-                  <div className="np-radio-progress-fill" style={{ width: `${Math.min(100, (radioMeta.elapsed / radioMeta.duration) * 100)}%` }} />
-                </div>
-                <span className="np-radio-time">{formatTime(radioMeta.duration)}</span>
-              </div>
-            )}
-          </div>
-        </div>
-        <div className="np-hero-cover-wrap">
-          {resolvedCover
-            ? <img src={resolvedCover} alt={currentRadio.name} className="np-cover" />
-            : radioMeta.currentArt
-              ? <img src={radioMeta.currentArt} alt="" className="np-cover" onError={e => { (e.target as HTMLImageElement).style.display = 'none'; }} />
-              : <div className="np-cover np-cover-fallback"><Cast size={52} /></div>}
-        </div>
-        <div style={{ flex: 1 }} />
-      </div>
-
-      {radioMeta.nextSong && (
-        <div className="np-info-card">
-          <div className="np-card-header">
-            <h3 className="np-card-title"><SkipForward size={13} style={{ marginRight: 5 }} />{t('radio.upNext')}</h3>
-          </div>
-          <div className="np-radio-next-track">
-            {radioMeta.nextSong.art && (
-              <img src={radioMeta.nextSong.art} alt="" className="np-radio-track-art"
-                onError={e => { (e.target as HTMLImageElement).style.display = 'none'; }} />
-            )}
-            <div className="np-radio-track-info">
-              <span className="np-radio-track-title">{radioMeta.nextSong.title}</span>
-              {radioMeta.nextSong.artist && <span className="np-radio-track-artist">{radioMeta.nextSong.artist}</span>}
-            </div>
-          </div>
-        </div>
-      )}
-
-      {radioMeta.history.length > 0 && (
-        <div className="np-info-card">
-          <div className="np-card-header">
-            <h3 className="np-card-title"><Clock size={13} style={{ marginRight: 5 }} />{t('radio.recentlyPlayed')}</h3>
-          </div>
-          <div className="np-album-tracklist">
-            {radioMeta.history.map((item, idx) => (
-              <div key={idx} className="np-album-track">
-                {item.song.art && (
-                  <img src={item.song.art} alt="" className="np-radio-track-art np-radio-track-art--sm"
-                    onError={e => { (e.target as HTMLImageElement).style.display = 'none'; }} />
-                )}
-                <span className="np-album-track-title truncate">
-                  {item.song.artist ? `${item.song.artist} — ${item.song.title}` : item.song.title}
-                </span>
-              </div>
-            ))}
-          </div>
-        </div>
       )}
     </div>
   );
